refactor mcp transform preparation
